refactor: playing music state
