+++ src/static/Components/Player.tsx
@@ -4,7 +4,7 @@ import { useMatch, useNavigate } from 'react-router-dom'
 import cls from 'classnames'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import { libraryState } from '../model/library'
-import { musicState, playListState} from '../model/music'
+import { musicState, playListState} from '../model/playing'
 import { TagEditer } from './TagEditer'
 import { PLAY_MODE } from '../consts'
 import { RESP_STATE } from '../../shareCommon/consts'
