+++ src/static/Pages/AlbumDetail.tsx
@@ -3,7 +3,7 @@ import * as style from './AlbumDetail.module.less'
 import { useParams, useNavigate } from 'react-router-dom'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import { albumDetailState } from '../model/albumDetail'
-import { musicState } from '../model/music'
+import { musicState } from '../model/playing'
 import { getAlbumDetail } from '../API'
 import Navibar from '../Components/Navibar'
 import Cover from '../Components/Cover'
