+++ src/static/Pages/Library.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useSetRecoilState, useRecoilValue } from 'recoil'
 import { libraryState, libraryScrollState } from '../model/library'
-import { musicState } from '../model/music'
+import { musicState } from '../model/playing'
 import { useLoadmore } from '../hooks/useLoadmore'
 import { getLibrary, scanLibrary } from '../API'
 import Songlist from '../Components/Songlist'
